Remove authentication until we know how to implement it exactly
