+++ apiserver/organization/organization.go
@@ -3,21 +3,15 @@ package organization
 import (
 	"context"
 	"fmt"
-	"time"
 
 	orgv1 "github.com/appuio/control-api/apis/organization/v1"
 
 	corev1 "k8s.io/api/core/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/api/meta"
 	metainternalversion "k8s.io/apimachinery/pkg/apis/meta/internalversion"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/selection"
-	"k8s.io/apimachinery/pkg/util/duration"
-	"k8s.io/apiserver/pkg/authorization/authorizer"
-	"k8s.io/apiserver/pkg/endpoints/filters"
 	genericregistry "k8s.io/apiserver/pkg/registry/generic"
 	"k8s.io/apiserver/pkg/registry/rest"
 	restbuilder "sigs.k8s.io/apiserver-runtime/pkg/builder/rest"
@@ -27,15 +21,13 @@ import (
 func New() restbuilder.ResourceHandlerProvider {
 	return func(s *runtime.Scheme, g genericregistry.RESTOptionsGetter) (rest.Storage, error) {
 		return &organizationStorage{
-			namepaces:           &loopbackNamespaceProvider{},
-			namespaceAuthorizer: &loopbackNamespaceAuthorizer{},
+			namepaces: &loopbackNamespaceProvider{},
 		}, nil
 	}
 }
 
 type organizationStorage struct {
-	namepaces           namespaceProvider
-	namespaceAuthorizer authorizer.Authorizer
+	namepaces namespaceProvider
 }
 type namespaceProvider interface {
 	getNamespace(ctx context.Context, name string, options *metav1.GetOptions) (*corev1.Namespace, error)
@@ -58,18 +50,6 @@ func (s *organizationStorage) NamespaceScoped() bool {
 var _ rest.Getter = &organizationStorage{}
 
 func (s *organizationStorage) Get(ctx context.Context, name string, options *metav1.GetOptions) (runtime.Object, error) {
-	res := &orgv1.Organization{}
-	a, err := filters.GetAuthorizerAttributes(ctx)
-	if err != nil {
-		return nil, err
-	}
-	decision, _, err := s.namespaceAuthorizer.Authorize(ctx, a)
-	if err != nil {
-		return nil, err
-	} else if decision != authorizer.DecisionAllow {
-		return nil, kerrors.NewNotFound(res.GetGroupVersionResource().GroupResource(), name)
-	}
-
 	nsName := orgNameToNamespaceName(name)
 	ns, err := s.namepaces.getNamespace(ctx, nsName, options)
 	if err != nil {
@@ -86,18 +66,6 @@ func (s *organizationStorage) Create(ctx context.Context, obj runtime.Object, cr
 	org, ok := obj.(*orgv1.Organization)
 	if !ok {
 		return nil, fmt.Errorf("not an organization: %#v", obj)
-	}
-
-	a, err := filters.GetAuthorizerAttributes(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	decision, _, err := s.namespaceAuthorizer.Authorize(ctx, a)
-	if err != nil {
-		return nil, err
-	} else if decision != authorizer.DecisionAllow {
-		return nil, kerrors.NewNotFound(org.GetGroupVersionResource().GroupResource(), org.Name)
 	}
 
 	// Validate Org
@@ -150,18 +118,6 @@ func (s *organizationStorage) Update(ctx context.Context, name string, objInfo r
 
 	newOrg := &orgv1.Organization{}
 
-	a, err := filters.GetAuthorizerAttributes(ctx)
-	if err != nil {
-		return nil, false, err
-	}
-
-	decision, _, err := s.namespaceAuthorizer.Authorize(ctx, a)
-	if err != nil {
-		return nil, false, err
-	} else if decision != authorizer.DecisionAllow {
-		return nil, false, kerrors.NewNotFound(newOrg.GetGroupVersionResource().GroupResource(), name)
-	}
-
 	oldOrg, err := s.Get(ctx, name, nil)
 	if err != nil {
 
@@ -205,51 +161,4 @@ func (s *organizationStorage) Delete(ctx context.Context, name string, deleteVal
 
 	ns, err := s.namepaces.deleteNamespace(ctx, orgNameToNamespaceName(name), options)
 	return namespaceToOrg(ns), false, err
-}
-
-func (s *organizationStorage) ConvertToTable(ctx context.Context, obj runtime.Object, tableOptions runtime.Object) (*metav1.Table, error) {
-	var table metav1.Table
-
-	orgs := []orgv1.Organization{}
-	if meta.IsListType(obj) {
-		orgList, ok := obj.(*orgv1.OrganizationList)
-		if !ok {
-			return nil, fmt.Errorf("not an organization: %#v", obj)
-		}
-		orgs = orgList.Items
-	} else {
-		org, ok := obj.(*orgv1.Organization)
-		if !ok {
-			return nil, fmt.Errorf("not an organization: %#v", obj)
-		}
-		orgs = append(orgs, *org)
-	}
-
-	for _, org := range orgs {
-		table.Rows = append(table.Rows, orgToTableRow(&org))
-	}
-
-	if opt, ok := tableOptions.(*metav1.TableOptions); !ok || !opt.NoHeaders {
-		desc := metav1.ObjectMeta{}.SwaggerDoc()
-		table.ColumnDefinitions = []metav1.TableColumnDefinition{
-			{Name: "Name", Type: "string", Format: "name", Description: desc["name"]},
-			{Name: "Display Name", Type: "string", Description: "Name of the organization"},
-			{Name: "Namespace", Type: "string", Description: "Name of the underlying namespace"},
-			{Name: "Age", Type: "date", Description: desc["creationTimestamp"]},
-		}
-	}
-	return &table, nil
-}
-
-func orgToTableRow(org *orgv1.Organization) metav1.TableRow {
-	nsName := ""
-	if org.Annotations != nil {
-		nsName = org.Annotations[namespaceKey]
-	}
-
-	return metav1.TableRow{
-		Cells:  []interface{}{org.GetName(), org.Spec.DisplayName, nsName, duration.HumanDuration(time.Since(org.GetCreationTimestamp().Time))},
-		Object: runtime.RawExtension{Object: org},
-	}
-
 }
